+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- added: Login info to testconfig for testpartners.ts
+- changed: Change Lifi and Thorchain DA to use variable quotes
+- fixed: LiFi and Thorchain/DA memo handling
+
 ## 2.3.1 (2024-03-29)
 
 - fixed: Use proper `EdgeMemo` API on `EdgeSpendInfo`